+++ src/Exception/Arrays/InvalidAdditionalTupleItemsException.php
@@ -7,11 +7,11 @@ namespace PHPModelGenerator\Exception\Arrays;
 use PHPModelGenerator\Exception\Generic\InvalidAdditionalPropertiesException;
 
 /**
- * Class InvalidAdditionalItemsException
+ * Class InvalidAdditionalTupleItemsException
  *
  * @package PHPModelGenerator\Exception\Arrays
  */
-class InvalidAdditionalItemsException extends InvalidAdditionalPropertiesException
+class InvalidAdditionalTupleItemsException extends InvalidAdditionalPropertiesException
 {
     protected const MAIN_MESSAGE = 'Tuple array %s contains invalid additional items.';
     protected const TYPE = 'additional item';
